requirements.txt
Appium-Python-Client
mitmproxy
requests
selenium
sqlalchemy